add back Default packages; move Dependency inside nix package
