+++ bobtask/task.go
@@ -2,6 +2,7 @@ package bobtask
 
 import (
 	"fmt"
+	"github.com/benchkram/bob/pkg/nix"
 	"path/filepath"
 	"strings"
 
@@ -101,7 +102,7 @@ type Task struct {
 	// dependencies contain the actual dependencies merged
 	// with the global dependencies defined in the Bobfile
 	// in the order which they need to be added to PATH
-	dependencies []Dependency
+	dependencies []nix.Dependency
 
 	// storePaths contain /nix/store/* paths
 	// in the order which they need to be added to PATH
@@ -122,7 +123,7 @@ func Make(opts ...TaskOption) Task {
 		env:                  []string{},
 		rebuild:              RebuildOnChange,
 		dockerRegistryClient: dockermobyutil.NewRegistryClient(),
-		dependencies:         []Dependency{},
+		dependencies:         []nix.Dependency{},
 	}
 
 	for _, opt := range opts {
@@ -180,10 +181,10 @@ func (t *Task) SetEnv(env []string) {
 	t.env = env
 }
 
-func (t *Task) Dependencies() []Dependency {
+func (t *Task) Dependencies() []nix.Dependency {
 	return t.dependencies
 }
-func (t *Task) SetDependencies(dependencies []Dependency) {
+func (t *Task) SetDependencies(dependencies []nix.Dependency) {
 	t.dependencies = dependencies
 }
 
@@ -377,24 +378,4 @@ func parseTargetImage(p string) []string {
 func keyExists(m map[string]interface{}, key string) bool {
 	_, ok := m[key]
 	return ok
-}
-
-type Dependency struct {
-	// Name of the dependency
-	Name string
-	// Nixpkgs can be empty or a link to desired revision ex. https://github.com/NixOS/nixpkgs/archive/eeefd01d4f630fcbab6588fe3e7fffe0690fbb20.tar.gz
-	Nixpkgs string
-}
-
-// UniqueDeps removes duplicates from the list by checking against name-nixpkgs key
-func UniqueDeps(s []Dependency) []Dependency {
-	added := make(map[string]bool)
-	var res []Dependency
-	for _, v := range s {
-		if _, exists := added[v.Name+v.Nixpkgs]; !exists {
-			res = append(res, v)
-			added[v.Name+v.Nixpkgs] = true
-		}
-	}
-	return res
 }
