+++ pkg/nix/default.go
@@ -1,11 +1,11 @@
 package nix
 
 // DefaultPackages which are installed with the rest of nix packages
-func DefaultPackages() []string {
-	return []string{
-		"bash",
-		"coreutils",
-		"gnused",
-		"findutils",
+func DefaultPackages() []Dependency {
+	return []Dependency{
+		{Name: "bash"},
+		{Name: "coreutils"},
+		{Name: "gnused"},
+		{Name: "findutils"},
 	}
 }
