+++ pkg/nix/dependency.go
@@ -1,0 +1,22 @@
+package nix
+
+type Dependency struct {
+	// Name of the dependency
+	Name string
+	// Nixpkgs can be empty or a link to desired revision
+	// ex. https://github.com/NixOS/nixpkgs/archive/eeefd01d4f630fcbab6588fe3e7fffe0690fbb20.tar.gz
+	Nixpkgs string
+}
+
+// UniqueDeps removes duplicates from the list by checking against name-nixpkgs key
+func UniqueDeps(s []Dependency) []Dependency {
+	added := make(map[string]bool)
+	var res []Dependency
+	for _, v := range s {
+		if _, exists := added[v.Name+v.Nixpkgs]; !exists {
+			res = append(res, v)
+			added[v.Name+v.Nixpkgs] = true
+		}
+	}
+	return res
+}
